7th commit
In here the user file is modify to auth
